+++ src/TradingApp.Module.Quotes/Application/Models/OscillationSettings.cs
@@ -8,10 +8,10 @@ public class OscillationSettings
     /// <summary>
     /// Gets or sets the threshold level below which an asset is considered oversold.
     /// </summary>
-    public double Oversold { get; set; }
+    public decimal Oversold { get; set; }
 
     /// <summary>
     /// Gets or sets the threshold level above which an asset is considered overbought.
     /// </summary>
-    public double Overbought { get; set; }
+    public decimal Overbought { get; set; }
 }
